Working user interface

+++ app/src/main/java/com/example/jackskitt/adlarcherydatalogger/Math/AHRS.java
@@ -214,12 +214,12 @@ public class AHRS {
 
     float invSqrt(float x) {
         float halfx = 0.5f * x;
-        float y = (float) x;
+        float y = x;
         //long i = *(long*)&y;
         int i = (int) y;
         i = 0x5f3759df - (i >> 1);
         //y = *(float*)&i;
-        y = (int) i;
+        y = i;
         y = y * (1.5f - (halfx * y * y));
         return y;
     }
